requirements.txt
PyOpenGL==3.1.0
numpy==1.10.1